daily-avatar/requirements.txt
google-genai
cnlunar>=0.2.4
Pillow>=10.0
requests>=2.31
